smarthomehub/requirements.txt
Flask
paho-mqtt
python-dotenv